requirements.txt
# Spark

pyspark==3.5.1

# Object storage

boto3==1.34.69

# API

fastapi==0.110.1
uvicorn[standard]==0.29.0

# Query engine (gold layer)

duckdb==0.10.2

# Dashboard

streamlit==1.33.0

# Data processing

pandas==2.2.1
pyarrow==15.0.2

# Utilities

python-dotenv==1.0.1
requests==2.31.0